+++ graphium/cli/__init__.py
@@ -1,1 +1,2 @@
-
+from .data import data_cli
+from .main import main_cli
